server ready

+++ index.js
@@ -48,21 +48,44 @@ async function run() {
         })
 
         // delete a book
-        app.delete('/book/:id', async (req, res) => {
-            const id = req.params.id;
-            const result = await Books.deleteOne({ _id: new ObjectId(id) })
-            res.send({ status: true, data: result })
+        app.delete('/book', async (req, res) => {
+            const id = req.query.id;
+            const email = req.query.user;
+            const selectedBook = await Books.findOne({ _id: new ObjectId(id) })
+            if (selectedBook.user === email) {
+                const result = await Books.deleteOne({ _id: new ObjectId(id) })
+                res.send({ status: true, data: result })
+            } else {
+                res.status(401).json({ status: false, message: 'you are not authorized to delete this book' })
+            }
         })
 
         // update a book
         app.patch('/book/:id', async (req, res) => {
             const id = req.params.id
+            const oldBook = await Books.findOne({ _id: new ObjectId(id) })
             const updateDoc = {
                 $set: req.body
             }
-            const result = await Books.updateOne({ _id: new ObjectId(id) }, updateDoc, { upsert: true })
+            if (oldBook.user === req.body.user) {
+                const result = await Books.updateOne({ _id: new ObjectId(id) }, updateDoc, { upsert: true })
+                res.send({ status: true, data: result })
+            } else {
+                res.status(401).json({ status: false, message: 'you are not authorized to edit this book' })
+            }
+        })
+
+
+        // leave a review
+        app.patch('/review/:id', async (req, res) => {
+            const id = req.params.id
+            const newReview = req.body.review
+            const book = await Books.findOne({ _id: new ObjectId(id) })
+            book.Reviews.push(newReview)
+            const result = await Books.updateOne({ _id: new ObjectId(id) }, { $set: book }, { upsert: true })
             res.send({ status: true, data: result })
         })
+
